Experiments as uv workspaces
