+++ .experiment_template/src/.amlignore
@@ -4,14 +4,8 @@
 # filenames to ignore in it. The .amlignore file uses
 # the same syntax and patterns as the .gitignore file.
 
-# This script
-.amlignore
-
 # Local scripts are personal testing ground for each dev
-local.py
-
-# Unit tests are irrelevant for AzureML jobs
-tests/
+debug.py
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
