+++ .experiment_template/src/azure-ml-job.yaml
@@ -1,0 +1,30 @@
+# Tells Azure ML what kind of YAML this is.
+#  Docs: https://docs.microsoft.com/en-us/azure/machine-learning/reference-yaml-job-command
+$schema: https://azuremlschemas.azureedge.net/latest/commandJob.schema.json
+
+# Name of the experiment where all jobs will end up in the Azure ML dashboard
+experiment_name: experiment_template
+
+# What to run
+command: >-
+  python -m experiment_template.main
+  --data_path "${{inputs.data_path}}"
+  --greeting "${{inputs.greeting}}"
+
+inputs:
+  data_path:
+    # Only works if dataset created using Azure ML CLI v2; run `az ml data create --help` to see how
+    #   YAML schema: https://learn.microsoft.com/en-us/azure/machine-learning/reference-yaml-data
+    type: uri_folder  # default, can be changed to `uri_file` if data_path points to a file
+    path: azureml:<azure-ml-dataset-name>:<azure-ml-dataset-version>
+  greeting: "Hello"
+
+# What code to make available
+code: .
+
+# Where to run it
+environment:
+  build:
+    path: ./environment
+
+compute: azureml:<azure-ml-cluster-name>
