+++ .experiment_template/src/debug_template.py
@@ -1,0 +1,12 @@
+"""
+This file should serve as a template on how to debug a similar payload as the
+one specified in the azure-ml-job.yaml. Users should copy it to have their own
+`debug.py`, which is gitignored so it can be modified at will.
+"""
+
+from pathlib import Path
+
+from experiment_template.main import main
+
+if __name__ == "__main__":
+    main(Path("an/example/path"), greeting="An example greeting")
