+++ .gitignore
@@ -1,5 +1,5 @@
 # Local entrypoints should be developer-specific
-**/local.py
+**/debug.py
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
