+++ .vscode/settings.json
@@ -1,0 +1,12 @@
+{
+    "python.defaultInterpreterPath": "${workspaceFolder}/.venv/bin/python",
+    "[python]": {"editor.defaultFormatter": "charliermarsh.ruff"},
+    "python.testing.pytestArgs": ["experiments"],
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true,
+    "editor.rulers": [
+      { "column": 80 },
+      { "column": 100 }
+    ],
+    "python.envFile": "${workspaceFolder}/container.env"
+}
