+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+After changes, the following workflow should work seamlessly:
+
+1. `bin/exp/new exp_test`
+2. Run local in VSCode.
+3. Run tests in VSCode.
+
+Removing an experiment
+
+TODO:
+run job in aml to test all works nice
