+++ bin/lint/py
@@ -4,7 +4,7 @@
 set -eo pipefail
 cd "$(dirname "$0")/../.."
 
-source bin/env .env
+source bin/env container.env
 bin/chkenv CODE_PATH
 
 # Usage documentation
