+++ container.env
@@ -6,7 +6,7 @@
 AZURE_ML_WORKSPACE="<azure-ml-workspace-name>"
 # Azure Machine Learning resource group scripts point to
 RESOURCE_GROUP="<azure-ml-resource-group>"
-# Relative folder where the source code lives in the project.
-CODE_PATH=./src
-# Relative folder where runs are isolated before submitting to AzureML
-ISOLATED_RUNS_PATH=./isolated_runs
+# Folder where the main code lives
+CODE_PATH=./experiments
+# Relative folder where runs are packaged before submitting to AzureML
+RUNS_PATH=./runs
