+++ experiments/example_experiment/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "example-experiment"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "azureml-mlflow==1.58.*",
+    "mlflow-skinny==2.18.*",
+    "pytz==2024.2",
+]
+
+[project.scripts]
+example-experiment = "example_experiment:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
