+++ experiments/example_experiment/src/example_experiment/__init__.py
@@ -1,0 +1,2 @@
+def main() -> None:
+    print("Hello from example-experiment!")
