+++ experiments/shared/pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "shared"
+version = "0.1.0"
+description = "Package shared across experiments"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = []
+
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
